Add experimental --enable-long-lnum configure option, with major edits.
This requires major cleanup in usage of in/cs_int_t/cs_lnum_t arguments.
At this stage, the code compiles without additional warnings, and has been
lightly tested on "mesh verification/qauliy mode". It is not compatible
with the Fortran API, as we do not use different integer kinds yet
(this would be feasible, but moving to C is the preferred solution).

+++ src/atmo/cs_at_data_assim.c
@@ -273,7 +273,7 @@ cs_at_data_assim_build_ops(void)
         bft_printf("    Obs %i\n",ii);
         for (cs_lnum_t jj = proj_idx[ii]; jj < proj_idx[ii+1]; jj++)
           bft_printf("    Point %i x %.2f y %.2f z %.2f coef %.2f\n",
-                     jj, (proj + jj*4)[1], (proj + jj*4)[2],
+                     (int)jj, (proj + jj*4)[1], (proj + jj*4)[2],
                      (proj + jj*4)[3], (proj + jj*4)[0]);
         bft_printf("\n");
       }
